requirements.txt
geoip2~=5.2.0
psutil~=7.1.3
httpx~=0.28.1
psutil~=7.1.3
pyyaml~=6.0.3